--- apps/proxy/qna/src/main/java/com/example/application/port/in/CreateInputPort.java
@@ -1,12 +1,0 @@
-package com.example.application.port.in;
-
-import com.example.framework.adapter.in.dto.in.CreateAnswerIn;
-import com.example.framework.adapter.in.dto.in.CreateQuestionIn;
-import com.example.framework.adapter.in.dto.out.CreateAnswerOut;
-import com.example.framework.adapter.in.dto.out.CreateQuestionOut;
-
-public interface CreateInputPort {
-    CreateQuestionOut createQuestion(CreateQuestionIn createQuestionIn);
-    CreateAnswerOut createAnswer(Long questionId, CreateAnswerIn createAnswerIn);
-    void addCoffeeMenu();
-}
