--- apps/proxy/qna/src/main/java/com/example/application/port/in/DeleteInputPort.java
@@ -1,9 +1,0 @@
-package com.example.application.port.in;
-
-import com.example.framework.adapter.in.dto.out.DeleteAnswerOut;
-import com.example.framework.adapter.in.dto.out.DeleteQuestionOut;
-
-public interface DeleteInputPort {
-    DeleteQuestionOut deleteQuestion(Long questionId);
-    DeleteAnswerOut deleteAnswer(Long answerId);
-}
