--- apps/proxy/qna/src/main/java/com/example/application/port/in/ModifyInputPort.java
@@ -1,5 +1,0 @@
-package com.example.application.port.in;
-
-public interface ModifyInputPort {
-    void modifyCoffeePrice();
-}
