+++ apps/proxy/qna/src/main/java/com/example/application/port/in/answer/CreateAnswerInputPort.java
@@ -1,0 +1,8 @@
+package com.example.application.port.in.answer;
+
+import com.example.framework.adapter.in.dto.in.CreateAnswerIn;
+import com.example.framework.adapter.in.dto.out.CreateAnswerOut;
+
+public interface CreateAnswerInputPort {
+    CreateAnswerOut createAnswer(Long questionId, CreateAnswerIn createAnswerIn);
+}
