+++ apps/proxy/qna/src/main/java/com/example/application/port/in/answer/DeleteAnswerInputPort.java
@@ -1,0 +1,7 @@
+package com.example.application.port.in.answer;
+
+import com.example.framework.adapter.in.dto.out.DeleteAnswerOut;
+
+public interface DeleteAnswerInputPort {
+    DeleteAnswerOut deleteAnswer(Long answerId);
+}
