+++ apps/proxy/qna/src/main/java/com/example/application/port/in/question/CreateQuestionInputPort.java
@@ -1,0 +1,8 @@
+package com.example.application.port.in.question;
+
+import com.example.framework.adapter.in.dto.in.CreateQuestionIn;
+import com.example.framework.adapter.in.dto.out.CreateQuestionOut;
+
+public interface CreateQuestionInputPort {
+    CreateQuestionOut createQuestion(CreateQuestionIn createQuestionIn);
+}
