+++ apps/proxy/qna/src/main/java/com/example/application/port/in/question/DeleteQuestionInputPort.java
@@ -1,0 +1,7 @@
+package com.example.application.port.in.question;
+
+import com.example.framework.adapter.in.dto.out.DeleteQuestionOut;
+
+public interface DeleteQuestionInputPort {
+    DeleteQuestionOut deleteQuestion(Long questionId);
+}
