--- apps/proxy/qna/src/main/java/com/example/application/service/ModifyUseCase.java
@@ -1,13 +1,0 @@
-package com.example.application.service;
-
-import com.example.application.port.in.ModifyInputPort;
-import org.springframework.stereotype.Service;
-
-@Service
-public class ModifyUseCase implements ModifyInputPort {
-
-    @Override
-    public void modifyCoffeePrice() {
-
-    }
-}
